+++ libs/flexible.js
@@ -1,12 +1,26 @@
+/**
+ * Required module
+ */
 const helpers = require('./helpers')
 
+/** Flexible Savings API Wrapper */
 module.exports = class Flexible {
+    /**
+     * Create a Flexible Savings Client
+     * @param {Object} binanceApiClient - Binance API Client
+     * @param {Array.<Object>} accountBalances - Account Balances
+     * @param {Array.<Object>} flexibleBalances - Flexible Savings Balances
+     */
     constructor(binanceApiClient, accountBalances, flexibleBalances) {
         this.binanceApiClient = binanceApiClient
         this.accountBalances = helpers.stripLendingPrefix(accountBalances)
         this.flexibleBalances = helpers.stripLendingPrefix(flexibleBalances)
     }
 
+    /**
+     * Get available Flexible Savings products
+     * @return {Array.<Object>} Available products
+     */
     async getAvailableProducts() {
         const products = await this.binanceApiClient.fetchApi(
             '/sapi/v1/lending/daily/product/list',
@@ -19,6 +33,10 @@ module.exports = class Flexible {
         )
     }
 
+    /**
+     * Purchase Flexible Savings products
+     * @param {Array.<Object>} products - Flexible Savings products to purchase
+     */
     async purchaseProducts(products) {
         await Promise.all(
             products.map(async product => {
@@ -31,11 +49,15 @@ module.exports = class Flexible {
                     { method: 'POST' }
                 )
 
-                helpers.log('FLEXIBLE purchase product', [product])
+                helpers.log('Purchase a Flexible Savings product', [product])
             })
         )
     }
 
+    /**
+     * Redeem Flexible Savings products
+     * @param {Array.<Object>} products - Flexible Savings products to redeem
+     */
     async redeemProducts(products) {
         await Promise.all(
             products.map(async product => {
@@ -56,7 +78,7 @@ module.exports = class Flexible {
                     { method: 'POST' }
                 )
 
-                helpers.log('FLEXIBLE redeem product', [product])
+                helpers.log('Redeem a Flexible Savings product', [product])
             })
         )
     }