+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Binance Savings Rebalancer / CHANGELOG
 
 
+### v1.1.0 (2020-12-04)
+* Add JSDoc documentation comments
+* Rename `Fixed Savings` to `Locked  Savings` similar to Binance
+
 ### v1.0.0 (2020-12-03)
-* Initial commit
+* Initial release